+++ packages/pangea-sdk/pangea/asyncio/services/authz.py
@@ -198,7 +198,7 @@ class AuthZAsync(ServiceBaseAsync):
     async def list_resources(
         self, type: str, action: str, subject: Subject, attributes: Optional[Dict[str, Any]] = None
     ) -> PangeaResponse[ListResourcesResult]:
-        """List resources. (Beta)
+        """List resources.
 
         Given a type, action, and subject, list all the resources in the
         type that the subject has access to the action with.
@@ -233,7 +233,7 @@ class AuthZAsync(ServiceBaseAsync):
     async def list_subjects(
         self, resource: Resource, action: str, attributes: Optional[Dict[str, Any]] = None
     ) -> PangeaResponse[ListSubjectsResult]:
-        """List subjects. (Beta)
+        """List subjects.
 
         Given a resource and an action, return the list of subjects who have
         access to the action for the given resource.