+++ packages/pangea-sdk/tests/integration/test_vault.py
@@ -105,7 +105,7 @@ class TestVault(unittest.TestCase):
         self.assertEqual(data_b64, decrypt_default_resp.result.plain_text)
 
         # Decrypt wrong id
-        with self.assertRaises(pe.ValidationException):
+        with self.assertRaises(pe.VaultItemNotFound):
             self.vault.decrypt("thisisnotandid", cipher_v2, 2)
 
         # Deactivate key
@@ -225,7 +225,7 @@ class TestVault(unittest.TestCase):
             self.vault.verify(id, data, signature_v2, 10)
 
         # Verify wrong id
-        with self.assertRaises(pe.ValidationException):
+        with self.assertRaises(pe.VaultItemNotFound):
             self.vault.verify("thisisnotandid", data, signature_v2, 2)
 
         # Verify wrong signature