+++ packages/pangea-sdk/pangea/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.3.0"
+__version__ = "4.4.0"
 
 from pangea.asyncio.request import PangeaRequestAsync
 from pangea.config import PangeaConfig