+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### 
+
+- Support for `domains` field in `v2/user/breached` endpoint in User Intel service
+
 ## 4.3.0 - 2024-09-25
 
 ### Added