+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### 
+### Added
 
 - Support for `domains` field in `v2/user/breached` endpoint in User Intel service