+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `domains` field in `v2/user/breached` endpoint in User Intel service
 - Detect-only Redact for Sanitize.
 
 ## 4.3.0 - 2024-09-25
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Dependency on the asyncio pypi package.
+- Lingering beta tags on AuthZ `/list-resources` and `/list-subjects` endpoints.
 
 ## 4.2.0 - 2024-07-16