+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 4.4.0 - 2024-10-15
+
 ### Added
 
 - Support for `domains` field in `v2/user/breached` endpoint in User Intel service